+++ collective/dancing/browser/subscribe.py
@@ -28,7 +28,7 @@ from collective.singing.channel import channel_lookup
 from collective.singing.interfaces import ISubscriptionKey
 from collective.dancing import MessageFactory as _
 from collective.dancing.utils import switch_on
-from collective.dancing.events import ConfirmSubscriptionEvent
+from collective.dancing.events import ConfirmSubscriptionEvent, ConfirmUnsubscriptionEvent
 from zope.event import notify
 
 
@@ -143,6 +143,7 @@ class Unsubscribe(BrowserView):
             if len(subscriptions):
                 for sub in subscriptions:
                     subs.remove_subscription(sub)
+                    notify(ConfirmUnsubscriptionEvent(self.context, sub))
                 self.status = _(u"You unsubscribed successfully.")
             else:
                 self.status = _(u"You aren't subscribed to this mailing-list.")
@@ -228,6 +229,7 @@ class SubscriptionEditForm(IncludeHiddenSecret, form.EditForm):
         subs = self.context.channel.subscriptions
         for subscription in subs.query(secret=secret):
             subs.remove_subscription(subscription)
+            notify(ConfirmUnsubscriptionEvent(subscription))
         self.removed = self.context
         self.status = _(u"You unsubscribed successfully.")
 
@@ -310,7 +312,6 @@ class SubscriptionAddForm(IncludeHiddenSecret, form.Form):
             self.request, zope.i18n.interfaces.IUserPreferredLanguages)
         if pl is not None:
             metadata['languages'] = pl.getPreferredLanguages()
-
         # By using another method here we allow subclasses to override
         # what really happens here:
         self.add_subscription(
@@ -322,6 +323,10 @@ class SubscriptionAddForm(IncludeHiddenSecret, form.Form):
         try:
             self.added = self.context.subscriptions.add_subscription(
                 self.context, secret, comp_data, coll_data, metadata)
+            if secret and not self.added.metadata.get('pending', False):
+                #in this case the user subscribed himself in my-subscriptions.html
+                #panel and he doesn't need to confirm his subscription
+                notify(ConfirmSubscriptionEvent(context, self.added))
         except ValueError:
             self.added = None
             self.status = self.status_already_subscribed
@@ -588,6 +593,7 @@ class SubscriptionEditSubForm(SubscriptionSubForm):
         subs = self.context.channel.subscriptions
         for subscription in subs.query(secret=secret, format=format):
             subs.remove_subscription(subscription)
+            notify(ConfirmUnsubscriptionEvent(self.context, subscription))
         self.removed = self.context
         self.status = self.status_unsubscribed